requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Medical imaging
pydicom>=2.3.0
SimpleITK>=2.1.0
vtk>=9.1.0
nibabel>=3.2.0

# Statistical analysis
statsmodels>=0.13.0
pingouin>=0.5.0

# Machine learning validation
imbalanced-learn>=0.8.0
scikit-image>=0.18.0

# Visualization and reporting
jupyter>=1.0.0
ipywidgets>=7.6.0
kaleido>=0.2.1
fpdf2>=2.5.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
nbsphinx>=0.8.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0
click>=8.0.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Development tools
black>=21.0.0
flake8>=4.0.0
mypy>=0.910

# Optional: GPU acceleration
# torch>=1.9.0
# torchvision>=0.10.0